+++ test/Altinn.App.Core.Tests/Helpers/JsonHelperTests.cs
@@ -240,16 +240,4 @@ public class JsonHelperTests
         diff.Should().ContainKey("ListTest[0].DecimalTest").WhoseValue.Should().Match(x => isMatch(x));
         diff.Should().HaveCount(5);
     }
-
-    [Fact]
-    public void OptionsWithIgnorePrefix()
-    {
-        var options1 = JsonHelper.GetOptionsWithIgnorePrefix("A_");
-        var options2 = JsonHelper.GetOptionsWithIgnorePrefix("A_");
-        options1.Should().BeSameAs(options2);
-        var options3 = JsonHelper.GetOptionsWithIgnorePrefix("B_");
-        options3.Should().NotBeSameAs(options2);
-
-        Assert.Throws<ArgumentNullException>(() => JsonHelper.GetOptionsWithIgnorePrefix(null!));
-    }
 }
